Improve eigen solver
