+++ src/symfc/basis_sets/basis_sets_O1.py
@@ -7,8 +7,8 @@ from typing import Optional
 import numpy as np
 from scipy.sparse import csr_array
 
+from symfc.eig_solvers.eig_tools import eigsh_projector
 from symfc.spg_reps import SpgRepsO1
-from symfc.utils.eig_tools import eigsh_projector
 from symfc.utils.translation_tools_O1 import compressed_projector_sum_rules
 from symfc.utils.utils import SymfcAtoms
 from symfc.utils.utils_O1 import (
