+++ src/symfc/basis_sets/basis_sets_O2.py
@@ -5,12 +5,12 @@ from __future__ import annotations
 import numpy as np
 from scipy.sparse import csr_array
 
-from symfc.spg_reps import SpgRepsO2
-from symfc.utils.eig_tools import (
+from symfc.eig_solvers.eig_tools import (
     eigsh_projector,
     eigsh_projector_sumrule,
 )
-from symfc.utils.matrix import BlockMatrixNode
+from symfc.eig_solvers.matrix import BlockMatrixNode
+from symfc.spg_reps import SpgRepsO2
 
 try:
     from symfc.utils.matrix import dot_product_sparse
