+++ src/symfc/basis_sets/basis_sets_O3.py
@@ -8,12 +8,12 @@ from typing import Optional, Union
 import numpy as np
 from scipy.sparse import coo_array, csr_array
 
-from symfc.spg_reps import SpgRepsO3
-from symfc.utils.eig_tools import (
+from symfc.eig_solvers.eig_tools import (
     eigsh_projector,
     eigsh_projector_sumrule,
 )
-from symfc.utils.matrix import BlockMatrixNode
+from symfc.eig_solvers.matrix import BlockMatrixNode
+from symfc.spg_reps import SpgRepsO3
 
 try:
     from symfc.utils.matrix import dot_product_sparse
