+++ src/symfc/eig_solvers/graph.py
